# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almáskosár API",
        "description": "Az Almáskosár Webshop API végpontjai",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listázza az összes felhasználót",
                "description": "Listázza az összes felhasználót",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Users not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Új felhasználó létrehozása",
                "description": "Új felhasználó létrehozása",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "requestBody": {
                    "description": "Új felhasználó adatai",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsersSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A felhasználó sikeresen létrejött",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Az üzenet",
                                            "type": "string",
                                            "example": "Sikeresen hozzáadva!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/edit": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Felhasználó keresése",
                "description": "Visszaadja a keresett felhasználó adatait",
                "operationId": "f85e77d28c01f3aa6de9ac3558344999",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Keresett felhasználó azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visszaadja a keresett felhasználót",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Felhasználó szerkesztése",
                "description": "Frissíti a keresett felhasználó adatait",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Frissíteni kívánt felhasználó azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsersSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Felhasználó törlése",
                "description": "Törli a keresett felhasználót",
                "operationId": "ce06860a51ab016ef66e753544fe4563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Törölni kívánt felhasználó azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Bejelentkezés email-el és jelszóval",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "A felhasználó hitelesítő adatainak átadása",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UsersSchema": {
                "properties": {
                    "id": {
                        "description": "Felhasználó azonosítója",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "name": {
                        "description": "Felhasználó neve",
                        "type": "string",
                        "example": "Jakab",
                        "nullable": true
                    },
                    "email": {
                        "description": "Felhasználó email címe",
                        "type": "string",
                        "format": "email",
                        "example": "gipszes.jakab@gmail.com",
                        "nullable": false
                    },
                    "email_verified_at": {
                        "description": "Felhasználó email cím ellenőrzésének ideje",
                        "type": "string",
                        "format": "timestamp",
                        "example": "2023-03-13 12:10:03",
                        "nullable": false
                    },
                    "password": {
                        "description": "Jelszó",
                        "type": "string",
                        "example": "************",
                        "nullable": false
                    },
                    "last_name": {
                        "description": "Felhasználó vezetékneve",
                        "type": "string",
                        "example": "Gipszes",
                        "nullable": true
                    },
                    "first_name": {
                        "description": "Felhasználó keresztneve",
                        "type": "string",
                        "example": "Jakab",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Felhasználó irányítószáma",
                        "type": "integer",
                        "example": "1454",
                        "nullable": true
                    },
                    "city": {
                        "description": "Felhasználó városa",
                        "type": "string",
                        "example": "Budapest",
                        "nullable": true
                    },
                    "address": {
                        "description": "Felhasználó címe",
                        "type": "string",
                        "example": "Lilaakác u. 15.",
                        "nullable": true
                    },
                    "phone_number": {
                        "description": "Felhasználó telefonszáma",
                        "type": "string",
                        "example": "+36302124787",
                        "nullable": true
                    },
                    "customer_code": {
                        "description": "Felhasználó vevőkódja",
                        "type": "string",
                        "example": "0",
                        "nullable": false
                    },
                    "admin": {
                        "description": "Felhasználó jogosultsága",
                        "type": "integer",
                        "example": "0",
                        "nullable": false
                    },
                    "remember_token": {
                        "description": "Token",
                        "type": "string",
                        "example": "V6SYqJB1Dp",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}